+++ src/stack_matrix_calculator.cc
@@ -14,15 +14,15 @@ const int MAX_SIZE = 100;
 
 
 void getDimensions(int matrixNumber, int dimensions[2]);
-void getMatrixValues(const int matrixNumber, int matrix[MAX_SIZE][MAX_SIZE],
+void getMatrixValues(const int matrixNumber, float matrix[MAX_SIZE][MAX_SIZE],
                      const int dimensions[2]);
-void printMatrix(const int matrixNumber, const int matrix[MAX_SIZE][MAX_SIZE],
+void printMatrix(const int matrixNumber, const float matrix[MAX_SIZE][MAX_SIZE],
                  const int dimensions[2]);
-void printSum(const int matrix1[MAX_SIZE][MAX_SIZE], const int matrix2[MAX_SIZE][MAX_SIZE],
+void printSum(const float matrix1[MAX_SIZE][MAX_SIZE], const float matrix2[MAX_SIZE][MAX_SIZE],
               const int dimensions1[2], const int dimensions2[2]);
-void printDifference(const int matrix1[MAX_SIZE][MAX_SIZE], const int matrix2[MAX_SIZE][MAX_SIZE],
+void printDifference(const float matrix1[MAX_SIZE][MAX_SIZE], const float matrix2[MAX_SIZE][MAX_SIZE],
                      const int dimensions1[2], const int dimensions2[2]);
-void printProduct(const int matrix1[MAX_SIZE][MAX_SIZE], const int matrix2[MAX_SIZE][MAX_SIZE],
+void printProduct(const float matrix1[MAX_SIZE][MAX_SIZE], const float matrix2[MAX_SIZE][MAX_SIZE],
                   const int dimensions1[2], const int dimensions2[2]);
 void printMenu();
 
@@ -36,8 +36,8 @@ int main(void) {
   getDimensions(1, dimensions1);
   getDimensions(2, dimensions2);
 
-  int matrix1[MAX_SIZE][MAX_SIZE];
-  int matrix2[MAX_SIZE][MAX_SIZE];
+  float matrix1[MAX_SIZE][MAX_SIZE];
+  float matrix2[MAX_SIZE][MAX_SIZE];
 
   // Get values for matrix 1.
   std::cout << std::endl;
@@ -135,7 +135,7 @@ void getDimensions(int matrixNumber, int dimensions[2]) {
  * @param matrix - the matrix to fill out with user input.
  * @param dimensions - the dimensions of the matrix.
  */
-void getMatrixValues(const int matrixNumber, int matrix[MAX_SIZE][MAX_SIZE],
+void getMatrixValues(const int matrixNumber, float matrix[MAX_SIZE][MAX_SIZE],
                      const int dimensions[2]) {
   std::cout << "===== Matrix " << matrixNumber << " =====" << std::endl;
   std::cout << "Enter " << dimensions[0] * dimensions[1] << " value(s) individually or seperated by space."
@@ -157,7 +157,7 @@ void getMatrixValues(const int matrixNumber, int matrix[MAX_SIZE][MAX_SIZE],
  * @param matrix - the matrix to print out.
  * @param dimensions - the dimensions of the matrix.
  */
-void printMatrix(const int matrixNumber, const int matrix[MAX_SIZE][MAX_SIZE],
+void printMatrix(const int matrixNumber, const float matrix[MAX_SIZE][MAX_SIZE],
                  const int dimensions[2]) {
   std::cout << "----- Matrix " << matrixNumber << " -----" << std::endl;
   // For every row.
@@ -195,7 +195,7 @@ void printMenu() {
  * @param dimensions1 - the dimensions of the first matrix.
  * @param dimensions2 - the dimensions of the second matrix.
  */
-void printSum(const int matrix1[MAX_SIZE][MAX_SIZE], const int matrix2[MAX_SIZE][MAX_SIZE],
+void printSum(const float matrix1[MAX_SIZE][MAX_SIZE], const float matrix2[MAX_SIZE][MAX_SIZE],
               const int dimensions1[2], const int dimensions2[2]) {
   // Check for same size.
   if (dimensions1[0] != dimensions2[0] || dimensions1[1] != dimensions2[1]) {
@@ -222,7 +222,7 @@ void printSum(const int matrix1[MAX_SIZE][MAX_SIZE], const int matrix2[MAX_SIZE]
  * @param dimensions1 - the dimensions of the first matrix.
  * @param dimensions2 - the dimensions of the second matrix.
  */
-void printDifference(const int matrix1[MAX_SIZE][MAX_SIZE], const int matrix2[MAX_SIZE][MAX_SIZE],
+void printDifference(const float matrix1[MAX_SIZE][MAX_SIZE], const float matrix2[MAX_SIZE][MAX_SIZE],
                      const int dimensions1[2], const int dimensions2[2]) {
   // Check for same size.
   if (dimensions1[0] != dimensions2[0] || dimensions1[1] != dimensions2[1]) {
@@ -249,7 +249,7 @@ void printDifference(const int matrix1[MAX_SIZE][MAX_SIZE], const int matrix2[MA
  * @param dimensions1 - the dimensions of the first matrix.
  * @param dimensions2 - the dimensions of the second matrix.
  */
-void printProduct(const int matrix1[MAX_SIZE][MAX_SIZE], const int matrix2[MAX_SIZE][MAX_SIZE],
+void printProduct(const float matrix1[MAX_SIZE][MAX_SIZE], const float matrix2[MAX_SIZE][MAX_SIZE],
                   const int dimensions1[2], const int dimensions2[2]) {
   // Check if matrix1's width == matrix2's height.
   if (dimensions1[0] != dimensions2[1]) {
@@ -262,7 +262,7 @@ void printProduct(const int matrix1[MAX_SIZE][MAX_SIZE], const int matrix2[MAX_S
   for (int i = 0; i < dimensions1[1]; ++i) {
     // For each column for matrix 2.
     for (int k = 0; k < dimensions2[0]; ++k) {
-      int sum = 0;
+      float sum = 0;
       // For each column in row for matrix 1.
       for (int j = 0; j < dimensions1[0]; ++j) {
         sum += matrix1[i][j] * matrix2[j][k];
